Create question variable create btn event listener
Create "question" variable that has the value of 16 and is changed
based on the user input.
When the user clicks the button to change the grid, they enter a new
value for the "question" variable. The button calls the createDivGrid()
function.

The function was changed to reset its contents, then enter a loop
that runs (question x question) times.
It creates the divs
adds the class 'grid-item' to them and adds an event listener to
each div.
That event listener calls the function highlight() and adds
a black background to the div item.

Then the div flex-basis value is changed so that the grid shows
based on the (question x question) ex. 16x16. It divides 100 by question
and returns a value.

Finaly it adds the div into the container.