Replace the grid adjusting button into a slider
Capture the slider from the html using its id #slider.
Capture the div that will contain the value of the slider.
Put the slider's value inside the variable sliderValue.
Create an event listener that detects to input in the slider.
When input is detected, change sliderValue into the slider's new value.
Change the contents of valueContainer to contain the slider's value.
Call the createDivGrid() function so that the grid is adjusted based on
the new input.
Adjust the createDivGrid() function, adding the sliderValue where
it needs to be.