Color the canvas only when the mouse is down
Previously, every single div on the grid had its own event listener
making it impossible to hold the mouse button and drag to draw on
the canvas.
Now, the grid container has the event listener that is activated by
mouse movement. when the mouse moves on the grid, the state of the
mouse being down is checked. if the mouse is down, the target,
or the element the mouse is moved on will be colored.